list_feature_classes(feature_dataset, feature_class)

list the feature classes available in ElmerGeo

Usage

list_feature_classes(feature_dataset = NULL, feature_class = NULL)

Arguments

feature_dataset

String. If supplied, this filters the list to just those feature classes in the feature dataset

feature_class

String. If supplied, this filters the list to just the named one. This can be useful if you know the name of the feature class but not the geography type or the feature dataset that it resides in.

Value

A dataframe listing the layer names along with the corresponding feature datasets and geometry types.. Defaults to NULL.

Examples

list_feature_classes()
#>                                     layer_name             feature_dataset
#> 1                 COUNTY_BACKGROUND_NOWATER_CP         branch_version_test
#> 2                       COUNTY_LINES_CP_BRANCH         branch_version_test
#> 3                      EQUITY_TRACTS_CP_BRANCH         branch_version_test
#> 4                               PSRC_REGION_CP         branch_version_test
#> 5                   PSRC_REGION_CPEAK_BRANCH_3         branch_version_test
#> 6                              PSRC_REGION_DH_         branch_version_test
#> 7                     PUBLIC_SCHOOLS_CP_BRANCH         branch_version_test
#> 8                                 REG20PUMA_CP         branch_version_test
#> 9         SCHOOL_ATTENDANCE_ZONES_CPEAK_BRANCH         branch_version_test
#> 10                             TAZ_BRANCH_TEST         branch_version_test
#> 11                                   BLOCK1980                      census
#> 12                                   BLOCK1990                      census
#> 13                                   BLOCK2000                      census
#> 14                           BLOCK2000_NOWATER                      census
#> 15                                   BLOCK2010                      census
#> 16                           BLOCK2010_NOWATER                      census
#> 17                                   BLOCK2020                      census
#> 18                           BLOCK2020_NOWATER                      census
#> 19                                BLOCKGRP1990                      census
#> 20                                BLOCKGRP2000                      census
#> 21                      BLOCKGRP2000_LAND_ONLY                      census
#> 22                                BLOCKGRP2010                      census
#> 23                        BLOCKGRP2010_NOWATER                      census
#> 24                                BLOCKGRP2020                      census
#> 25                        BLOCKGRP2020_NOWATER                      census
#> 26                           CENSUS_PLACES2020                      census
#> 27                                   PLACE1990                      census
#> 28                                   PLACE2000                      census
#> 29                                   PLACE2010                      census
#> 30                                   PLACE2020                      census
#> 31                                   REG00PUMA                      census
#> 32                                   REG10PUMA                      census
#> 33                                   REG20PUMA                      census
#> 34                                   REG90PUMA                      census
#> 35                                 SUPERPUMA00                      census
#> 36                                     TAD2010                      census
#> 37                                   TRACT1940                      census
#> 38                                   TRACT1950                      census
#> 39                                   TRACT1960                      census
#> 40                                   TRACT1970                      census
#> 41                                   TRACT1980                      census
#> 42                                   TRACT1990                      census
#> 43                                   TRACT2000                      census
#> 44                           TRACT2000_NOWATER                      census
#> 45                                   TRACT2010                      census
#> 46                           TRACT2010_NOWATER                      census
#> 47                                   TRACT2020                      census
#> 48                           TRACT2020_NOWATER                      census
#> 49                                UZA2010_FHWA                      census
#> 50                                 UZA2010_FTA                      census
#> 51                                UZA2020_FHWA                      census
#> 52                                 UZA2020_FTA                      census
#> 53                                      FAULTS                 environment
#> 54                                 FLOODPLAINS                 environment
#> 55                                 GEO_HAZARDS                 environment
#> 56                         LARGEST_WATERBODIES                 environment
#> 57                                     MJR_WTR                 environment
#> 58                                NAT_RESOURCE                 environment
#> 59                             NATIONAL_FOREST                 environment
#> 60                               NATIONAL_PARK                 environment
#> 61                          RESOURCE_LANDS_ALL                 environment
#> 62                                      RIVERS                 environment
#> 63                                STEEP_SLOPES                 environment
#> 64                                 WATERBODIES                 environment
#> 65                                    WETLANDS                 environment
#> 66                                        WRIA                 environment
#> 67                           DISPLACEMENT_2018                      equity
#> 68                           DISPLACEMENT_RISK                      equity
#> 69                      DISPLACEMENT_RISK_MASK                      equity
#> 70                               EQUITY_TRACTS                      equity
#> 71                            OPPORTUNITY_MASK                      equity
#> 72                             REDLINING_AREAS                      equity
#> 73                          CITIES18_DASHBOARD                    forecast
#> 74                     CONTROL_HCT18_DASHBOARD                    forecast
#> 75                         CONTROL18_DASHBOARD                    forecast
#> 76                                    FAZ_2010                    forecast
#> 77                           FAZ_LRG_AREAS2000                    forecast
#> 78                           FAZ_LRG_AREAS2010                    forecast
#> 79                                     FAZ1991                    forecast
#> 80                                     FAZ2000                    forecast
#> 81                                 FAZ2000_GRP                    forecast
#> 82                            FAZ2010_NO_WATER                    forecast
#> 83                                         FLU                    forecast
#> 84                                    REGION17                    forecast
#> 85                         SOUNDCAST_DISTRICTS                    forecast
#> 86                     SOUNDCAST_TAZ_DISTRICTS                    forecast
#> 87                         SUBREGS18_DASHBOARD                    forecast
#> 88                          TARGET18_DASHBOARD                    forecast
#> 89                      TAZ_TRACT_AREA_WEIGHTS                    forecast
#> 90                                     TAZ2000                    forecast
#> 91                           TAZ2000_LAND_AREA                    forecast
#> 92                            TAZ2000_NO_WATER                    forecast
#> 93                             TAZ2000_UPDATED                    forecast
#> 94                     TAZ2000_UPDATED_NOWATER                    forecast
#> 95                                     TAZ2010                    forecast
#> 96                            TAZ2010_NO_WATER                    forecast
#> 97                                TAZ2010_TEST                    forecast
#> 98                          TAZ2010_TO_TAD2010                    forecast
#> 99                                       TAZ91                    forecast
#> 100                       URBANSIM_GRID_POINTS                    forecast
#> 101            PSRC_ASSESSOR_HOUSING_ESTIMATES                     housing
#> 102                                   SUBAREAS                     housing
#> 103                               TRANSITLINES              model_networks
#> 104                              TRANSREFEDGES              model_networks
#> 105                            AQUATIC_SYSTEMS                  open_space
#> 106         CONSERVATION_NEEDS_AQUATIC_SYSTEMS                  open_space
#> 107                CONSERVATION_NEEDS_FARMLAND                  open_space
#> 108           CONSERVATION_NEEDS_NATURAL_LANDS                  open_space
#> 109         CONSERVATION_NEEDS_WORKING_FORESTS                  open_space
#> 110                                   FARMLAND                  open_space
#> 111       NATIONAL_FOREST_LAND_NOT_HARVESTABLE                  open_space
#> 112                              NATURAL_LANDS                  open_space
#> 113                           OPEN_SPACE_PARKS                  open_space
#> 114                            WORKING_FORESTS                  open_space
#> 115                            BLDG_FOOTPRINTS                     parcels
#> 116                      PARCELS_URBANSIM_2014                     parcels
#> 117                  PARCELS_URBANSIM_2014_PTS                     parcels
#> 118                      PARCELS_URBANSIM_2018                     parcels
#> 119                  PARCELS_URBANSIM_2018_PTS                     parcels
#> 120                                       PLSS                     parcels
#> 121                          REG_MAJ_PUB_LANDS                     parcels
#> 122                                     CITIES                   political
#> 123                              CITIES_POINTS                   political
#> 124                                 CONGDIST12                   political
#> 125                          COUNTY_BACKGROUND                   political
#> 126                 COUNTY_BACKGROUND_NO_WATER                   political
#> 127                               COUNTY_LINES                   political
#> 128                          FEDERAL_URBAN_AID                   political
#> 129                    HIGHER_EDUCATION_POINTS                   political
#> 130                                      MICEN                   political
#> 131                               MICEN_POINTS                   political
#> 132                             MILITARY_BASES                   political
#> 133                      MILITARY_BASES_POINTS                   political
#> 134                                PSRC_REGION                   political
#> 135                      PSRC_REGIONAL_OUTLINE                   political
#> 136                                       PTBA                   political
#> 137                             PUBLIC_SCHOOLS                   political
#> 138                       REGIONAL_GEOGRAPHIES                   political
#> 139                                     RTPOWA                   political
#> 140                                      RURAL                   political
#> 141                    SCHOOL_ATTENDANCE_ZONES                   political
#> 142                           SCHOOL_DISTRICTS                   political
#> 143                  SCHOOL_DISTRICTS_NO_WATER                   political
#> 144                     SOUND_TRANSIT_DISTRICT                   political
#> 145                   TRIBAL_ENTERPRISE_POINTS                   political
#> 146                                TRIBAL_LAND                   political
#> 147                            TRIBAL_PROPERTY                   political
#> 148           UNINCORPORATED_URBAN_GROWTH_AREA                   political
#> 149                              URBAN_CENTERS                   political
#> 150                       URBAN_CENTERS_POINTS                   political
#> 151                          URBAN_GROWTH_AREA                   political
#> 152                     URBAN_GROWTH_AREA_LINE                   political
#> 153                             URBAN_VILLAGES                   political
#> 154                                WA_COUNTIES                   political
#> 155                  WA_COUNTIES_WATER_REMOVED                   political
#> 156                   WA_LEGISLATIVE_DISTRICTS                   political
#> 157                                  ZIP_CODES                   political
#> 158                          HEX_90_ACRES_CLIP                      raster
#> 159                         HEX_GRID_180_ACRES                      raster
#> 160                          HEX_GRID_90_ACRES                      raster
#> 161               TRANSIT_ROUTE_STOPS_2018_RTP                    rtp_2022
#> 162               TRANSIT_ROUTE_STOPS_2050_RTP                    rtp_2022
#> 163                    TRANSIT_ROUTES_2018_RTP                    rtp_2022
#> 164                    TRANSIT_ROUTES_2050_RTP                    rtp_2022
#> 165                     TRANSIT_STOPS_2018_RTP                    rtp_2022
#> 166                     TRANSIT_STOPS_2050_RTP                    rtp_2022
#> 167                               PARCELS_TEST                        test
#> 168                   SOUNDCAST_DISTRICTS_TEST                        test
#> 169                                  CITIES_CP           test_dataset_2_cp
#> 170                                   TAZ_TEST                       test2
#> 171                             AIRPORTS_WSDOT          transportation_air
#> 172                             MAJOR_AIRPORTS          transportation_air
#> 173                               REG_AIRPORTS          transportation_air
#> 174                           DEEP_WATER_PORTS      transportation_freight
#> 175                                     FGTSWA      transportation_freight
#> 176                 FREIGHT_ECONOMIC_CORRIDORS      transportation_freight
#> 177              FREIGHT_INTERMODAL_FACILITIES      transportation_freight
#> 178            FREIGHT_TRUCK_ECONOMIC_CORRIDOR      transportation_freight
#> 179                     FREIGHT_WATERWAYS_FGTS      transportation_freight
#> 180                    GROSS_SUPPLY_V2_2014_10      transportation_freight
#> 181                          HIGHWAY_USDOT_MFN      transportation_freight
#> 182                      INTERMODAL_CONNECTORS      transportation_freight
#> 183               INTERMODAL_CONNECTORS_LINEAR      transportation_freight
#> 184               INTERMODAL_CONNECTORS_POINTS      transportation_freight
#> 185                             MARINE_HIGHWAY      transportation_freight
#> 186                                  MFN_PORTS      transportation_freight
#> 187               NHFN_CRITICAL_URBAN_RURAL_FC      transportation_freight
#> 188                                  NHFN_PSRC      transportation_freight
#> 189                   OLYMPIC_PIPELINE_DRAWING      transportation_freight
#> 190                                      PORTS      transportation_freight
#> 191                         PRIVATE_TRUCK_STOP      transportation_freight
#> 192                           SAFETY_REST_AREA      transportation_freight
#> 193                                   STRAHNET      transportation_freight
#> 194                              WATER_FREIGHT      transportation_freight
#> 195                              WEIGH_STATION      transportation_freight
#> 196                                DRAFT_RBN17 transportation_nonmotorized
#> 197                           REGIONAL_BIKEPED transportation_nonmotorized
#> 198                              REGIONAL_SUPS transportation_nonmotorized
#> 199                         AG_CROSSINGS_STEP1         transportation_rail
#> 200                         AG_CROSSINGS_STEP2         transportation_rail
#> 201                                  PASS_RAIL         transportation_rail
#> 202                                       RAIL         transportation_rail
#> 203                                  RAIL_FGTS         transportation_rail
#> 204                            RAIL_FGTS_STATE         transportation_rail
#> 205                                   RAILWAYS         transportation_rail
#> 206                         BRIDGES_LOCAL_2020     transportation_roadways
#> 207                         BRIDGES_STATE_2020     transportation_roadways
#> 208                          FISH_PASSAGE_2020     transportation_roadways
#> 209                          FUNC_CLASS_NON_SR     transportation_roadways
#> 210                              FUNC_CLASS_SR     transportation_roadways
#> 211                                   HIGHWAYS     transportation_roadways
#> 212                         HIN_LIMITED_ACCESS     transportation_roadways
#> 213                        HIN_SURFACE_STREETS     transportation_roadways
#> 214                                        HSS     transportation_roadways
#> 215                                INTERSTATES     transportation_roadways
#> 216                                ITS_SIGNALS     transportation_roadways
#> 217                             MJR_HWYS_STATE     transportation_roadways
#> 218                                        MTS     transportation_roadways
#> 219                          NAT_HWY_SYS_LOCAL     transportation_roadways
#> 220                          NAT_HWY_SYS_STATE     transportation_roadways
#> 221                                        NHS     transportation_roadways
#> 222                             PARK_AND_RIDES     transportation_roadways
#> 223                                   REG_STCL     transportation_roadways
#> 224                        SR_MILE_POST_MARKER     transportation_roadways
#> 225                               STATE_ROUTES     transportation_roadways
#> 226                                        T_1     transportation_roadways
#> 227                                        T_2     transportation_roadways
#> 228                 CITY_TDM_PROGRAMS_07162020          transportation_tdm
#> 229                      CTR_WORK_SITES_REGION          transportation_tdm
#> 230                          COMBINED_PSRC_ADA      transportation_transit
#> 231                          HCT_STATION_AREAS      transportation_transit
#> 232                 HCT_STATION_AREAS_DISSOLVE      transportation_transit
#> 233                                  HCT_STOPS      transportation_transit
#> 234                                  HCT_TRANS      transportation_transit
#> 235                          TRANSIT_DISTRICTS      transportation_transit
#> 236                              TRANSIT_LINES      transportation_transit
#> 237                TRANSIT_NETWORK_ROUTES_2017      transportation_transit
#> 238                TRANSIT_NETWORK_ROUTES_2025      transportation_transit
#> 239                TRANSIT_NETWORK_ROUTES_2040      transportation_transit
#> 240                 TRANSIT_NETWORK_STOPS_2017      transportation_transit
#> 241                 TRANSIT_NETWORK_STOPS_2025      transportation_transit
#> 242                 TRANSIT_NETWORK_STOPS_2040      transportation_transit
#> 243                                  HCT_BLOCK                  vision2050
#> 244                             HCT_VISION_PTS                  vision2050
#> 245                           HCT_VISION_TRANS                  vision2050
#> 246                                LOCAL_PARKS                  vision2050
#> 247              LOCK_PARKS_QUARTERMILE_BUFFER                  vision2050
#> 248 REGIONAL_GEOGRAPHIES_PREFERRED_ALTERNATIVE                  vision2050
#> 249                               TOD_PRCL_UGA                  vision2050
#> 250                             TRANSIT_BUFFER                  vision2050
#> 251                UGA_QUARTER_BUFFER_NO_WATER                  vision2050
#> 252                   VISION_HCT_STATION_AREAS                  vision2050
#> 253         VISION_HCT_STATION_AREAS_DISSOLVED                  vision2050
#> 254     VISION_HCT_STATION_AREAS_RGC_DISSOLVED                  vision2050
#>            geometry_type
#> 1    esriGeometryPolygon
#> 2   esriGeometryPolyline
#> 3    esriGeometryPolygon
#> 4    esriGeometryPolygon
#> 5    esriGeometryPolygon
#> 6    esriGeometryPolygon
#> 7      esriGeometryPoint
#> 8    esriGeometryPolygon
#> 9    esriGeometryPolygon
#> 10   esriGeometryPolygon
#> 11   esriGeometryPolygon
#> 12   esriGeometryPolygon
#> 13   esriGeometryPolygon
#> 14   esriGeometryPolygon
#> 15   esriGeometryPolygon
#> 16   esriGeometryPolygon
#> 17   esriGeometryPolygon
#> 18   esriGeometryPolygon
#> 19   esriGeometryPolygon
#> 20   esriGeometryPolygon
#> 21   esriGeometryPolygon
#> 22   esriGeometryPolygon
#> 23   esriGeometryPolygon
#> 24   esriGeometryPolygon
#> 25   esriGeometryPolygon
#> 26   esriGeometryPolygon
#> 27   esriGeometryPolygon
#> 28   esriGeometryPolygon
#> 29   esriGeometryPolygon
#> 30   esriGeometryPolygon
#> 31   esriGeometryPolygon
#> 32   esriGeometryPolygon
#> 33   esriGeometryPolygon
#> 34   esriGeometryPolygon
#> 35   esriGeometryPolygon
#> 36   esriGeometryPolygon
#> 37   esriGeometryPolygon
#> 38   esriGeometryPolygon
#> 39   esriGeometryPolygon
#> 40   esriGeometryPolygon
#> 41   esriGeometryPolygon
#> 42   esriGeometryPolygon
#> 43   esriGeometryPolygon
#> 44   esriGeometryPolygon
#> 45   esriGeometryPolygon
#> 46   esriGeometryPolygon
#> 47   esriGeometryPolygon
#> 48   esriGeometryPolygon
#> 49   esriGeometryPolygon
#> 50   esriGeometryPolygon
#> 51   esriGeometryPolygon
#> 52   esriGeometryPolygon
#> 53  esriGeometryPolyline
#> 54   esriGeometryPolygon
#> 55   esriGeometryPolygon
#> 56   esriGeometryPolygon
#> 57   esriGeometryPolygon
#> 58   esriGeometryPolygon
#> 59   esriGeometryPolygon
#> 60   esriGeometryPolygon
#> 61   esriGeometryPolygon
#> 62  esriGeometryPolyline
#> 63   esriGeometryPolygon
#> 64   esriGeometryPolygon
#> 65   esriGeometryPolygon
#> 66   esriGeometryPolygon
#> 67   esriGeometryPolygon
#> 68   esriGeometryPolygon
#> 69   esriGeometryPolygon
#> 70   esriGeometryPolygon
#> 71   esriGeometryPolygon
#> 72   esriGeometryPolygon
#> 73   esriGeometryPolygon
#> 74   esriGeometryPolygon
#> 75   esriGeometryPolygon
#> 76   esriGeometryPolygon
#> 77   esriGeometryPolygon
#> 78   esriGeometryPolygon
#> 79   esriGeometryPolygon
#> 80   esriGeometryPolygon
#> 81   esriGeometryPolygon
#> 82   esriGeometryPolygon
#> 83   esriGeometryPolygon
#> 84   esriGeometryPolygon
#> 85   esriGeometryPolygon
#> 86   esriGeometryPolygon
#> 87   esriGeometryPolygon
#> 88   esriGeometryPolygon
#> 89   esriGeometryPolygon
#> 90   esriGeometryPolygon
#> 91   esriGeometryPolygon
#> 92   esriGeometryPolygon
#> 93   esriGeometryPolygon
#> 94   esriGeometryPolygon
#> 95   esriGeometryPolygon
#> 96   esriGeometryPolygon
#> 97   esriGeometryPolygon
#> 98   esriGeometryPolygon
#> 99   esriGeometryPolygon
#> 100    esriGeometryPoint
#> 101    esriGeometryPoint
#> 102  esriGeometryPolygon
#> 103 esriGeometryPolyline
#> 104 esriGeometryPolyline
#> 105  esriGeometryPolygon
#> 106  esriGeometryPolygon
#> 107  esriGeometryPolygon
#> 108  esriGeometryPolygon
#> 109  esriGeometryPolygon
#> 110  esriGeometryPolygon
#> 111  esriGeometryPolygon
#> 112  esriGeometryPolygon
#> 113  esriGeometryPolygon
#> 114  esriGeometryPolygon
#> 115  esriGeometryPolygon
#> 116  esriGeometryPolygon
#> 117    esriGeometryPoint
#> 118  esriGeometryPolygon
#> 119    esriGeometryPoint
#> 120  esriGeometryPolygon
#> 121  esriGeometryPolygon
#> 122  esriGeometryPolygon
#> 123    esriGeometryPoint
#> 124  esriGeometryPolygon
#> 125  esriGeometryPolygon
#> 126  esriGeometryPolygon
#> 127 esriGeometryPolyline
#> 128  esriGeometryPolygon
#> 129    esriGeometryPoint
#> 130  esriGeometryPolygon
#> 131    esriGeometryPoint
#> 132  esriGeometryPolygon
#> 133    esriGeometryPoint
#> 134  esriGeometryPolygon
#> 135  esriGeometryPolygon
#> 136  esriGeometryPolygon
#> 137    esriGeometryPoint
#> 138  esriGeometryPolygon
#> 139  esriGeometryPolygon
#> 140  esriGeometryPolygon
#> 141  esriGeometryPolygon
#> 142  esriGeometryPolygon
#> 143  esriGeometryPolygon
#> 144  esriGeometryPolygon
#> 145    esriGeometryPoint
#> 146  esriGeometryPolygon
#> 147  esriGeometryPolygon
#> 148  esriGeometryPolygon
#> 149  esriGeometryPolygon
#> 150    esriGeometryPoint
#> 151  esriGeometryPolygon
#> 152 esriGeometryPolyline
#> 153  esriGeometryPolygon
#> 154  esriGeometryPolygon
#> 155  esriGeometryPolygon
#> 156  esriGeometryPolygon
#> 157  esriGeometryPolygon
#> 158  esriGeometryPolygon
#> 159  esriGeometryPolygon
#> 160  esriGeometryPolygon
#> 161    esriGeometryPoint
#> 162    esriGeometryPoint
#> 163 esriGeometryPolyline
#> 164 esriGeometryPolyline
#> 165    esriGeometryPoint
#> 166    esriGeometryPoint
#> 167  esriGeometryPolygon
#> 168  esriGeometryPolygon
#> 169  esriGeometryPolygon
#> 170  esriGeometryPolygon
#> 171    esriGeometryPoint
#> 172    esriGeometryPoint
#> 173    esriGeometryPoint
#> 174    esriGeometryPoint
#> 175 esriGeometryPolyline
#> 176 esriGeometryPolyline
#> 177    esriGeometryPoint
#> 178 esriGeometryPolyline
#> 179 esriGeometryPolyline
#> 180  esriGeometryPolygon
#> 181 esriGeometryPolyline
#> 182    esriGeometryPoint
#> 183 esriGeometryPolyline
#> 184    esriGeometryPoint
#> 185 esriGeometryPolyline
#> 186    esriGeometryPoint
#> 187 esriGeometryPolyline
#> 188 esriGeometryPolyline
#> 189 esriGeometryPolyline
#> 190    esriGeometryPoint
#> 191    esriGeometryPoint
#> 192    esriGeometryPoint
#> 193 esriGeometryPolyline
#> 194 esriGeometryPolyline
#> 195    esriGeometryPoint
#> 196 esriGeometryPolyline
#> 197 esriGeometryPolyline
#> 198 esriGeometryPolyline
#> 199    esriGeometryPoint
#> 200    esriGeometryPoint
#> 201 esriGeometryPolyline
#> 202 esriGeometryPolyline
#> 203 esriGeometryPolyline
#> 204 esriGeometryPolyline
#> 205 esriGeometryPolyline
#> 206    esriGeometryPoint
#> 207    esriGeometryPoint
#> 208    esriGeometryPoint
#> 209 esriGeometryPolyline
#> 210 esriGeometryPolyline
#> 211 esriGeometryPolyline
#> 212 esriGeometryPolyline
#> 213 esriGeometryPolyline
#> 214 esriGeometryPolyline
#> 215 esriGeometryPolyline
#> 216    esriGeometryPoint
#> 217 esriGeometryPolyline
#> 218 esriGeometryPolyline
#> 219 esriGeometryPolyline
#> 220 esriGeometryPolyline
#> 221 esriGeometryPolyline
#> 222    esriGeometryPoint
#> 223 esriGeometryPolyline
#> 224    esriGeometryPoint
#> 225 esriGeometryPolyline
#> 226 esriGeometryPolyline
#> 227 esriGeometryPolyline
#> 228  esriGeometryPolygon
#> 229    esriGeometryPoint
#> 230  esriGeometryPolygon
#> 231  esriGeometryPolygon
#> 232  esriGeometryPolygon
#> 233    esriGeometryPoint
#> 234 esriGeometryPolyline
#> 235  esriGeometryPolygon
#> 236 esriGeometryPolyline
#> 237 esriGeometryPolyline
#> 238 esriGeometryPolyline
#> 239 esriGeometryPolyline
#> 240    esriGeometryPoint
#> 241    esriGeometryPoint
#> 242    esriGeometryPoint
#> 243  esriGeometryPolygon
#> 244    esriGeometryPoint
#> 245 esriGeometryPolyline
#> 246  esriGeometryPolygon
#> 247  esriGeometryPolygon
#> 248  esriGeometryPolygon
#> 249  esriGeometryPolygon
#> 250  esriGeometryPolygon
#> 251  esriGeometryPolygon
#> 252  esriGeometryPolygon
#> 253  esriGeometryPolygon
#> 254  esriGeometryPolygon
list_feature_classes(feature_dataset = "equity")
#>               layer_name feature_dataset       geometry_type
#> 1      DISPLACEMENT_2018          equity esriGeometryPolygon
#> 2      DISPLACEMENT_RISK          equity esriGeometryPolygon
#> 3 DISPLACEMENT_RISK_MASK          equity esriGeometryPolygon
#> 4          EQUITY_TRACTS          equity esriGeometryPolygon
#> 5       OPPORTUNITY_MASK          equity esriGeometryPolygon
#> 6        REDLINING_AREAS          equity esriGeometryPolygon