list_recordsets(schema_name, recordset_type)

return a list of tables and views available in PSRC's in-house data warehouse

Usage

list_recordsets(schema_name = "", include_base_tables = FALSE)

Arguments

schema_name

(string, optional) If supplied, this limits the returned list to only those tables and views in schema_name.

include_base_tables

(TRUE|FALSE, optional) If false, return only the views (no tables). False is the default.

Value

a list of tables, in the form of a dataframe

Examples

list_recordsets()
#>                  schema
#> 1         park_and_ride
#> 2           small_areas
#> 3              HHSurvey
#> 4                equity
#> 5              HHSurvey
#> 6                equity
#> 7                equity
#> 8              HHSurvey
#> 9              HHSurvey
#> 10             HHSurvey
#> 11        regional_gtfs
#> 12        regional_gtfs
#> 13          data_portal
#> 14        regional_gtfs
#> 15             HHSurvey
#> 16          data_portal
#> 17        regional_gtfs
#> 18          data_portal
#> 19        regional_gtfs
#> 20               census
#> 21        regional_gtfs
#> 22             HHSurvey
#> 23        regional_gtfs
#> 24               census
#> 25 employment_summaries
#> 26               census
#> 27 employment_summaries
#> 28               census
#> 29               census
#> 30 employment_summaries
#> 31                  ofm
#> 32                  ofm
#> 33           employment
#> 34                 irhd
#> 35        park_and_ride
#> 36                  ofm
#> 37            Political
#> 38    parking_inventory
#> 39          data_portal
#> 40              general
#> 41          small_areas
#> 42          data_portal
#> 43          small_areas
#> 44          small_areas
#> 45               census
#> 46          small_areas
#> 47          small_areas
#> 48          small_areas
#> 49          small_areas
#> 50            ofm_views
#> 51          small_areas
#> 52       seatac_airport
#> 53             HHSurvey
#> 54        Macroeconomic
#> 55        Macroeconomic
#> 56        Macroeconomic
#> 57        Macroeconomic
#> 58                 chas
#> 59               equity
#> 60             HHSurvey
#> 61           employment
#> 62        park_and_ride
#> 63    parking_inventory
#> 64               equity
#> 65               census
#> 66               equity
#> 67                  LUV
#> 68          small_areas
#> 69               equity
#> 70          small_areas
#> 71          small_areas
#> 72                  LUV
#> 73           employment
#> 74          small_areas
#> 75          small_areas
#> 76                  LUV
#> 77          small_areas
#> 78                  LUV
#> 79          small_areas
#> 80                  LUV
#> 81                  LUV
#> 82                  LUV
#> 83               equity
#> 84          data_portal
#> 85             HHSurvey
#> 86             HHSurvey
#>                                              recordset_name recordset_type
#> 1                                 v_park_and_ride_dashboard           VIEW
#> 2                                            v_parcels_2018           VIEW
#> 3                                          v_value_metadata           VIEW
#> 4                                       v_blockgroup_shares           VIEW
#> 5                                       v_households_labels           VIEW
#> 6                                      v_tracker_indicators           VIEW
#> 7                                            v_tract_shares           VIEW
#> 8                                          v_persons_labels           VIEW
#> 9                                            v_trips_labels           VIEW
#> 10                                                v_persons           VIEW
#> 11                                            v_agency_2018           VIEW
#> 12                                          v_calendar_2018           VIEW
#> 13                         v_assessor_net_housing_by_county           VIEW
#> 14                                            v_routes_2018           VIEW
#> 15                                            v_days_labels           VIEW
#> 16                   v_assessor_net_housing_by_jurisdiction           VIEW
#> 17                                        v_service_id_2018           VIEW
#> 18                          v_assessor_net_housing_by_tract           VIEW
#> 19                                        v_stop_times_2018           VIEW
#> 20                                          v_geography_dim           VIEW
#> 21                                             v_stops_2018           VIEW
#> 22                                        v_vehicles_labels           VIEW
#> 23                                             v_trips_2018           VIEW
#> 24                               v_data_for_member_profiles           VIEW
#> 25                   v_covered_employment_by_city_by_naics2           VIEW
#> 26                               v_geo_relationships_blocks           VIEW
#> 27             v_covered_employment_by_city_by_major_sector           VIEW
#> 28                         v_geo_relationships_block_groups           VIEW
#> 29                               v_geo_relationships_tracts           VIEW
#> 30            v_covered_employment_by_tract_by_major_sector           VIEW
#> 31                                         v_estimates_2019           VIEW
#> 32                                v_current_block_estimates           VIEW
#> 33                                                test_view           VIEW
#> 34                                   v_data_years_available           VIEW
#> 35                                 v_park_and_ride_modeling           VIEW
#> 36                                         v_estimates_2020           VIEW
#> 37                                   incorporated_city_dims           VIEW
#> 38                                 v_parking_inventory_view           VIEW
#> 39                                         v_park_and_rides           VIEW
#> 40                               parcel_level_census_splits           VIEW
#> 41                                      v_luvit_2023_county           VIEW
#> 42                                 v_parking_by_block_group           VIEW
#> 43                                        v_luvit_2023_city           VIEW
#> 44                                v_luvit_2023_target_geogs           VIEW
#> 45                                      v_pums_persons_2019           VIEW
#> 46                               v_luvit_2023_control_geogs           VIEW
#> 47                           v_luvit_2023_control_hct_geogs           VIEW
#> 48                                         v_luvit_2023_faz           VIEW
#> 49                                         v_luvit_2023_taz           VIEW
#> 50                                         v_estimates_2019           VIEW
#> 51                                     v_luvit_2023_tract10           VIEW
#> 52                                           v_measurements           VIEW
#> 53                              v_column_definitions_public           VIEW
#> 54                                             v_employment           VIEW
#> 55                                             v_households           VIEW
#> 56                                                 v_income           VIEW
#> 57                                                    v_pop           VIEW
#> 58 v_households_by_cost_burden_by_tenure_by_race_by_subarea           VIEW
#> 59                                      v_blockgroup_counts           VIEW
#> 60                                             v_households           VIEW
#> 61                           v_hct_station_areas_employment           VIEW
#> 62                                         v_park_and_rides           VIEW
#> 63                                              v_test_view           VIEW
#> 64                                    v_tracker_indicators2           VIEW
#> 65  v_pums_households_by_rent_cost_burden_by_income_by_race           VIEW
#> 66                                    v_tracker_indicators3           VIEW
#> 67                               v_jobs_by_sector_by_county           VIEW
#> 68                                        v_household_facts           VIEW
#> 69                                    v_tracker_indicators4           VIEW
#> 70                               v_jobs_by_sector_by_county           VIEW
#> 71                                              v_job_facts           VIEW
#> 72                                                   v_jobs           VIEW
#> 73                                         custom_geography           VIEW
#> 74                                                   v_jobs           VIEW
#> 75                                            v_parcel_dims           VIEW
#> 76                                        v_household_facts           VIEW
#> 77                                       v_population_facts           VIEW
#> 78                                              v_job_facts           VIEW
#> 79                                            v_sector_dims           VIEW
#> 80                                            v_parcel_dims           VIEW
#> 81                                       v_population_facts           VIEW
#> 82                                            v_sector_dims           VIEW
#> 83                                           v_tract_counts           VIEW
#> 84                                              v_test_view           VIEW
#> 85                                                   v_days           VIEW
#> 86                                               v_vehicles           VIEW

list_recordsets(schema_name="HHSurvey")
#>      schema              recordset_name recordset_type
#> 1  HHSurvey v_column_definitions_public           VIEW
#> 2  HHSurvey                      v_days           VIEW
#> 3  HHSurvey               v_days_labels           VIEW
#> 4  HHSurvey                v_households           VIEW
#> 5  HHSurvey         v_households_labels           VIEW
#> 6  HHSurvey                   v_persons           VIEW
#> 7  HHSurvey            v_persons_labels           VIEW
#> 8  HHSurvey              v_trips_labels           VIEW
#> 9  HHSurvey            v_value_metadata           VIEW
#> 10 HHSurvey                  v_vehicles           VIEW
#> 11 HHSurvey           v_vehicles_labels           VIEW

list_recordsets(schema_name="HHSurvey", include_base_tables=TRUE)
#>      schema              recordset_name recordset_type
#> 1  HHSurvey               day_fact_2023     BASE TABLE
#> 2  HHSurvey              days_2017_2019     BASE TABLE
#> 3  HHSurvey                   days_2021     BASE TABLE
#> 4  HHSurvey                   days_2023     BASE TABLE
#> 5  HHSurvey             hh_survey_audit     BASE TABLE
#> 6  HHSurvey          household_dim_2014     BASE TABLE
#> 7  HHSurvey          household_dim_2017     BASE TABLE
#> 8  HHSurvey         household_fact_2015     BASE TABLE
#> 9  HHSurvey         household_fact_2023     BASE TABLE
#> 10 HHSurvey        households_2017_2019     BASE TABLE
#> 11 HHSurvey             households_2021     BASE TABLE
#> 12 HHSurvey             households_2023     BASE TABLE
#> 13 HHSurvey             person_dim_2014     BASE TABLE
#> 14 HHSurvey             person_dim_2015     BASE TABLE
#> 15 HHSurvey            person_fact_2014     BASE TABLE
#> 16 HHSurvey            person_fact_2015     BASE TABLE
#> 17 HHSurvey            person_fact_2023     BASE TABLE
#> 18 HHSurvey           persons_2017_2019     BASE TABLE
#> 19 HHSurvey                persons_2021     BASE TABLE
#> 20 HHSurvey                persons_2023     BASE TABLE
#> 21 HHSurvey                        Trip     BASE TABLE
#> 22 HHSurvey               trip_dim_2014     BASE TABLE
#> 23 HHSurvey               trip_dim_2015     BASE TABLE
#> 24 HHSurvey              trip_fact_2014     BASE TABLE
#> 25 HHSurvey              trip_fact_2023     BASE TABLE
#> 26 HHSurvey             trips_2017_2019     BASE TABLE
#> 27 HHSurvey                  trips_2021     BASE TABLE
#> 28 HHSurvey                  trips_2023     BASE TABLE
#> 29 HHSurvey v_column_definitions_public           VIEW
#> 30 HHSurvey                      v_days           VIEW
#> 31 HHSurvey               v_days_labels           VIEW
#> 32 HHSurvey                v_households           VIEW
#> 33 HHSurvey         v_households_labels           VIEW
#> 34 HHSurvey                   v_persons           VIEW
#> 35 HHSurvey            v_persons_labels           VIEW
#> 36 HHSurvey              v_trips_labels           VIEW
#> 37 HHSurvey            v_value_metadata           VIEW
#> 38 HHSurvey                  v_vehicles           VIEW
#> 39 HHSurvey           v_vehicles_labels           VIEW
#> 40 HHSurvey              value_metadata     BASE TABLE
#> 41 HHSurvey                  values_dim     BASE TABLE
#> 42 HHSurvey           variable_metadata     BASE TABLE
#> 43 HHSurvey          variable_metadata2     BASE TABLE
#> 44 HHSurvey          variables_codebook     BASE TABLE
#> 45 HHSurvey           vehicle_fact_2023     BASE TABLE
#> 46 HHSurvey          vehicles_2017_2019     BASE TABLE
#> 47 HHSurvey               vehicles_2021     BASE TABLE
#> 48 HHSurvey               vehicles_2023     BASE TABLE