list_feature_classes(feature_dataset, feature_class)

list the feature classes available in ElmerGeo

Usage

list_feature_classes(feature_dataset = NULL, feature_class = NULL)

Arguments

feature_dataset

String. If supplied, this filters the list to just those feature classes in the feature dataset

feature_class

String. If supplied, this filters the list to just the named one. This can be useful if you know the name of the feature class but not the geography type or the feature dataset that it resides in.

Value

A dataframe listing the layer names along with the corresponding feature datasets and geometry types.. Defaults to NULL.

Examples

list_feature_classes()
#>                                     layer_name             feature_dataset
#> 1                                    BLOCK1980                      census
#> 2                                    BLOCK1990                      census
#> 3                                    BLOCK2000                      census
#> 4                               BLOCK2000_CLIP                      census
#> 5                            BLOCK2000_NOWATER                      census
#> 6                                    BLOCK2010                      census
#> 7                            BLOCK2010_NOWATER                      census
#> 8                       BLOCK2010_SPATIAL_JOIN                      census
#> 9                                    BLOCK2020                      census
#> 10                           BLOCK2020_NOWATER                      census
#> 11                                BLOCKGRP1990                      census
#> 12                                BLOCKGRP2000                      census
#> 13                      BLOCKGRP2000_LAND_ONLY                      census
#> 14                                BLOCKGRP2010                      census
#> 15                        BLOCKGRP2010_NOWATER                      census
#> 16                                BLOCKGRP2020                      census
#> 17                        BLOCKGRP2020_NOWATER                      census
#> 18                               BREMERTON_UZA                      census
#> 19                               CENSUS_PLACES                      census
#> 20                     HWY_URBANIZED_AREA_2013                      census
#> 21                              MARYSVILLE_UZA                      census
#> 22                                   PLACE1990                      census
#> 23                                   PLACE2000                      census
#> 24                                   PLACE2010                      census
#> 25                                   PLACE2020                      census
#> 26                                   REG00PUMA                      census
#> 27                                   REG10PUMA                      census
#> 28                                   REG20PUMA                      census
#> 29                                   REG90PUMA                      census
#> 30                                     STE_UZA                      census
#> 31                                 SUPERPUMA00                      census
#> 32                                     TAD2010                      census
#> 33                                   TRACT1940                      census
#> 34                                   TRACT1950                      census
#> 35                                   TRACT1960                      census
#> 36                                   TRACT1970                      census
#> 37                                   TRACT1980                      census
#> 38                                   TRACT1990                      census
#> 39                                   TRACT2000                      census
#> 40                           TRACT2000_NOWATER                      census
#> 41                                   TRACT2010                      census
#> 42                           TRACT2010_NOWATER                      census
#> 43                                   TRACT2020                      census
#> 44                           TRACT2020_NOWATER                      census
#> 45                                      UA2010                      census
#> 46                                UZA2010_FHWA                      census
#> 47                                 UZA2010_FTA                      census
#> 48                                      AGLAND                 environment
#> 49                                      FAULTS                 environment
#> 50                                 FLOODPLAINS                 environment
#> 51                                 GEO_HAZARDS                 environment
#> 52                         LARGEST_WATERBODIES                 environment
#> 53                                     MJR_WTR                 environment
#> 54                                NAT_RESOURCE                 environment
#> 55                             NATIONAL_FOREST                 environment
#> 56                               NATIONAL_PARK                 environment
#> 57                                  NWI_REGION                 environment
#> 58                                      PARKOS                 environment
#> 59                              PUGET_SOUND_AA                 environment
#> 60                          RESOURCE_LANDS_ALL                 environment
#> 61                                      RIVERS                 environment
#> 62                                STEEP_SLOPES                 environment
#> 63                        URBANSIM_ENV_OVERLAY                 environment
#> 64                                 WATERBODIES                 environment
#> 65                                    WETLANDS                 environment
#> 66                                     WETLKIN                 environment
#> 67                                     WETLKIT                 environment
#> 68                                     WETLPIE                 environment
#> 69                                     WETLSNO                 environment
#> 70                                   WETLSNO03                 environment
#> 71                                        WRIA                 environment
#> 72                           DISPLACEMENT_2018                      equity
#> 73                           DISPLACEMENT_RISK                      equity
#> 74                      DISPLACEMENT_RISK_MASK                      equity
#> 75                               EQUITY_TRACTS                      equity
#> 76                            OPPORTUNITY_MASK                      equity
#> 77                             REDLINING_AREAS                      equity
#> 78                          CITIES18_DASHBOARD                    forecast
#> 79                         CONTROL18_DASHBOARD                    forecast
#> 80                                    FAZ_2010                    forecast
#> 81                           FAZ_LRG_AREAS2000                    forecast
#> 82                           FAZ_LRG_AREAS2010                    forecast
#> 83                                     FAZ1991                    forecast
#> 84                                     FAZ2000                    forecast
#> 85                                 FAZ2000_GRP                    forecast
#> 86                            FAZ2010_NO_WATER                    forecast
#> 87                                    REGION17                    forecast
#> 88                                      SDST00                    forecast
#> 89                                   SKAGITTAZ                    forecast
#> 90                         SOUNDCAST_DISTRICTS                    forecast
#> 91                     SOUNDCAST_TAZ_DISTRICTS                    forecast
#> 92                         SUBREGS18_DASHBOARD                    forecast
#> 93                          TARGET18_DASHBOARD                    forecast
#> 94                      TAZ_TRACT_AREA_WEIGHTS                    forecast
#> 95                                     TAZ2000                    forecast
#> 96                              TAZ2000_ISSKTH                    forecast
#> 97                           TAZ2000_LAND_AREA                    forecast
#> 98                            TAZ2000_NO_WATER                    forecast
#> 99                             TAZ2000_UPDATED                    forecast
#> 100                    TAZ2000_UPDATED_NOWATER                    forecast
#> 101                                    TAZ2010                    forecast
#> 102                           TAZ2010_NO_WATER                    forecast
#> 103                         TAZ2010_TO_TAD2010                    forecast
#> 104                                      TAZ91                    forecast
#> 105                       URBANSIM_GRID_POINTS                    forecast
#> 106                                   SUBAREAS                     housing
#> 107                               TRANSITLINES              model_networks
#> 108                              TRANSREFEDGES              model_networks
#> 109                                     AQPM10                  monitoring
#> 110                                  PM25_HARN                  monitoring
#> 111                            AQUATIC_SYSTEMS                  open_space
#> 112         CONSERVATION_NEEDS_AQUATIC_SYSTEMS                  open_space
#> 113                CONSERVATION_NEEDS_FARMLAND                  open_space
#> 114           CONSERVATION_NEEDS_NATURAL_LANDS                  open_space
#> 115         CONSERVATION_NEEDS_WORKING_FORESTS                  open_space
#> 116                                   FARMLAND                  open_space
#> 117       NATIONAL_FOREST_LAND_NOT_HARVESTABLE                  open_space
#> 118                              NATURAL_LANDS                  open_space
#> 119                           OPEN_SPACE_PARKS                  open_space
#> 120                            WORKING_FORESTS                  open_space
#> 121                            BLDG_FOOTPRINTS                     parcels
#> 122                               KING_PARCELS                     parcels
#> 123                           KITSAP_PARCEL_08                     parcels
#> 124                             KITSAP_PARCELS                     parcels
#> 125                      PARCELS_URBANSIM_2014                     parcels
#> 126                  PARCELS_URBANSIM_2014_PTS                     parcels
#> 127                      PARCELS_URBANSIM_2018                     parcels
#> 128                  PARCELS_URBANSIM_2018_PTS                     parcels
#> 129                                    PARKOS_                     parcels
#> 130                             PIERCE_PARCELS                     parcels
#> 131                         PIERCE_PARCELS_SPS                     parcels
#> 132                                       PLSS                     parcels
#> 133                          REG_MAJ_PUB_LANDS                     parcels
#> 134                                  SITE_ADDR                     parcels
#> 135                          SNOHOMISH_PARCELS                     parcels
#> 136                                     CITIES                   political
#> 137                              CITIES_POINTS                   political
#> 138                                 CONGDIST12                   political
#> 139                          COUNTY_BACKGROUND                   political
#> 140                               COUNTY_LINES                   political
#> 141                          FEDERAL_URBAN_AID                   political
#> 142                    HIGHER_EDUCATION_POINTS                   political
#> 143                                K12_SCHOOLS                   political
#> 144                                      MICEN                   political
#> 145                 MICEN_ARLINGTON_MARYSVILLE                   political
#> 146                               MICEN_POINTS                   political
#> 147                             MICEN_PROPOSED                   political
#> 148                             MILITARY_BASES                   political
#> 149                      MILITARY_BASES_POINTS                   political
#> 150                                PSRC_REGION                   political
#> 151                      PSRC_REGIONAL_OUTLINE                   political
#> 152                                       PTBA                   political
#> 153                       REGIONAL_GEOGRAPHIES                   political
#> 154                                     RTPOWA                   political
#> 155                                      RURAL                   political
#> 156                           SCHOOL_DISTRICTS                   political
#> 157                               SCHOOLS_K_12                   political
#> 158                     SOUND_TRANSIT_DISTRICT                   political
#> 159                   TRIBAL_ENTERPRISE_POINTS                   political
#> 160                                TRIBAL_LAND                   political
#> 161                            TRIBAL_PROPERTY                   political
#> 162           UNINCORPORATED_URBAN_GROWTH_AREA                   political
#> 163                              URBAN_CENTERS                   political
#> 164                       URBAN_CENTERS_POINTS                   political
#> 165                          URBAN_GROWTH_AREA                   political
#> 166                     URBAN_GROWTH_AREA_LINE                   political
#> 167                             URBAN_VILLAGES                   political
#> 168                                WA_COUNTIES                   political
#> 169                  WA_COUNTIES_WATER_REMOVED                   political
#> 170                   WA_LEGISLATIVE_DISTRICTS                   political
#> 171                                  ZIP_CODES                   political
#> 172                               CMP_SUBAREAS                    projects
#> 173                                        FLU                    projects
#> 174                          HEX_90_ACRES_CLIP                      raster
#> 175                         HEX_GRID_180_ACRES                      raster
#> 176                          HEX_GRID_90_ACRES                      raster
#> 177               TRANSIT_ROUTE_STOPS_2018_RTP                    rtp_2022
#> 178               TRANSIT_ROUTE_STOPS_2050_RTP                    rtp_2022
#> 179                    TRANSIT_ROUTES_2018_RTP                    rtp_2022
#> 180                    TRANSIT_ROUTES_2050_RTP                    rtp_2022
#> 181                     TRANSIT_STOPS_2018_RTP                    rtp_2022
#> 182                     TRANSIT_STOPS_2050_RTP                    rtp_2022
#> 183                  CENSUS_BLOCK_GROUPS_2020_                        test
#> 184                        CENSUS_BLOCKS_2020_                        test
#> 185                  CENSUS_BLOCKS_2020_POINTS                        test
#> 186                                    CITIES_                        test
#> 187                         CITIES_SINGLE_PART                        test
#> 188                                 EMME_LINKS                        test
#> 189                                      FLU_1                        test
#> 190                               FULL_PARCELS                        test
#> 191                        KING_COUNTY_PARCELS                        test
#> 192                 KING_COUNTY_PARCELS_POINTS                        test
#> 193                     PARCELS_URBANSIM_2018_                        test
#> 194                 PARCELS_URBANSIM_2018_PTS_                        test
#> 195                              PROJECTROUTES                        test
#> 196                        REG10PUMA_TEST_COPY                        test
#> 197                           SELECTED_PARCELS                        test
#> 198                                   TAZ_TEST                        test
#> 199                         TEST_HOSTED_BLOCKS                        test
#> 200               TEST_VERSIONING_MOVE_TO_BASE                        test
#> 201                             TRANSREFEDGES_                        test
#> 202                             AIRPORTS_WSDOT          transportation_air
#> 203                             MAJOR_AIRPORTS          transportation_air
#> 204                               REG_AIRPORTS          transportation_air
#> 205                           DEEP_WATER_PORTS      transportation_freight
#> 206                                     FGTSWA      transportation_freight
#> 207                 FREIGHT_ECONOMIC_CORRIDORS      transportation_freight
#> 208              FREIGHT_INTERMODAL_FACILITIES      transportation_freight
#> 209            FREIGHT_TRUCK_ECONOMIC_CORRIDOR      transportation_freight
#> 210                     FREIGHT_WATERWAYS_FGTS      transportation_freight
#> 211                    GROSS_SUPPLY_V2_2014_10      transportation_freight
#> 212                          HIGHWAY_USDOT_MFN      transportation_freight
#> 213                      INTERMODAL_CONNECTORS      transportation_freight
#> 214               INTERMODAL_CONNECTORS_LINEAR      transportation_freight
#> 215               INTERMODAL_CONNECTORS_POINTS      transportation_freight
#> 216                             MARINE_HIGHWAY      transportation_freight
#> 217                                  MFN_PORTS      transportation_freight
#> 218               NHFN_CRITICAL_URBAN_RURAL_FC      transportation_freight
#> 219                                  NHFN_PSRC      transportation_freight
#> 220                   OLYMPIC_PIPELINE_DRAWING      transportation_freight
#> 221                                      PORTS      transportation_freight
#> 222                         PRIVATE_TRUCK_STOP      transportation_freight
#> 223                           SAFETY_REST_AREA      transportation_freight
#> 224                                   STRAHNET      transportation_freight
#> 225                              WATER_FREIGHT      transportation_freight
#> 226                              WEIGH_STATION      transportation_freight
#> 227                                   WHS_WKPL      transportation_freight
#> 228                                   WHS_WKPT      transportation_freight
#> 229                                DRAFT_RBN17 transportation_nonmotorized
#> 230                           REGIONAL_BIKEPED transportation_nonmotorized
#> 231                              REGIONAL_SUPS transportation_nonmotorized
#> 232                         AG_CROSSINGS_STEP1         transportation_rail
#> 233                         AG_CROSSINGS_STEP2         transportation_rail
#> 234                                  PASS_RAIL         transportation_rail
#> 235                                       RAIL         transportation_rail
#> 236                                  RAIL_FGTS         transportation_rail
#> 237                            RAIL_FGTS_STATE         transportation_rail
#> 238                                   RAILWAYS         transportation_rail
#> 239                         BRIDGES_LOCAL_2020     transportation_roadways
#> 240                         BRIDGES_STATE_2020     transportation_roadways
#> 241                          FISH_PASSAGE_2020     transportation_roadways
#> 242                          FUNC_CLASS_NON_SR     transportation_roadways
#> 243                              FUNC_CLASS_SR     transportation_roadways
#> 244                                   HIGHWAYS     transportation_roadways
#> 245                                        HSS     transportation_roadways
#> 246                                INTERSTATES     transportation_roadways
#> 247                                ITS_SIGNALS     transportation_roadways
#> 248                             MJR_HWYS_STATE     transportation_roadways
#> 249                                        MTS     transportation_roadways
#> 250                          NAT_HWY_SYS_LOCAL     transportation_roadways
#> 251                          NAT_HWY_SYS_STATE     transportation_roadways
#> 252                                        NHS     transportation_roadways
#> 253                             PARK_AND_RIDES     transportation_roadways
#> 254                                   REG_STCL     transportation_roadways
#> 255                        SR_MILE_POST_MARKER     transportation_roadways
#> 256                               STATE_ROUTES     transportation_roadways
#> 257                                        T_1     transportation_roadways
#> 258                                        T_2     transportation_roadways
#> 259                 CITY_TDM_PROGRAMS_07162020          transportation_tdm
#> 260                      CTR_WORK_SITES_REGION          transportation_tdm
#> 261                          COMBINED_PSRC_ADA      transportation_transit
#> 262                          HCT_STATION_AREAS      transportation_transit
#> 263                 HCT_STATION_AREAS_DISSOLVE      transportation_transit
#> 264                                  HCT_STOPS      transportation_transit
#> 265                                  HCT_TRANS      transportation_transit
#> 266                          TRANSIT_DISTRICTS      transportation_transit
#> 267                              TRANSIT_LINES      transportation_transit
#> 268                TRANSIT_NETWORK_ROUTES_2017      transportation_transit
#> 269                TRANSIT_NETWORK_ROUTES_2025      transportation_transit
#> 270                TRANSIT_NETWORK_ROUTES_2040      transportation_transit
#> 271                 TRANSIT_NETWORK_STOPS_2017      transportation_transit
#> 272                 TRANSIT_NETWORK_STOPS_2025      transportation_transit
#> 273                 TRANSIT_NETWORK_STOPS_2040      transportation_transit
#> 274                                  HCT_BLOCK                  vision2050
#> 275                             HCT_VISION_PTS                  vision2050
#> 276                           HCT_VISION_TRANS                  vision2050
#> 277                                LOCAL_PARKS                  vision2050
#> 278              LOCK_PARKS_QUARTERMILE_BUFFER                  vision2050
#> 279 REGIONAL_GEOGRAPHIES_PREFERRED_ALTERNATIVE                  vision2050
#> 280                               TOD_PRCL_UGA                  vision2050
#> 281                             TRANSIT_BUFFER                  vision2050
#> 282                UGA_QUARTER_BUFFER_NO_WATER                  vision2050
#> 283                   VISION_HCT_STATION_AREAS                  vision2050
#> 284         VISION_HCT_STATION_AREAS_DISSOLVED                  vision2050
#> 285     VISION_HCT_STATION_AREAS_RGC_DISSOLVED                  vision2050
#>            geometry_type
#> 1    esriGeometryPolygon
#> 2    esriGeometryPolygon
#> 3    esriGeometryPolygon
#> 4    esriGeometryPolygon
#> 5    esriGeometryPolygon
#> 6    esriGeometryPolygon
#> 7    esriGeometryPolygon
#> 8    esriGeometryPolygon
#> 9    esriGeometryPolygon
#> 10   esriGeometryPolygon
#> 11   esriGeometryPolygon
#> 12   esriGeometryPolygon
#> 13   esriGeometryPolygon
#> 14   esriGeometryPolygon
#> 15   esriGeometryPolygon
#> 16   esriGeometryPolygon
#> 17   esriGeometryPolygon
#> 18   esriGeometryPolygon
#> 19   esriGeometryPolygon
#> 20   esriGeometryPolygon
#> 21   esriGeometryPolygon
#> 22   esriGeometryPolygon
#> 23   esriGeometryPolygon
#> 24   esriGeometryPolygon
#> 25   esriGeometryPolygon
#> 26   esriGeometryPolygon
#> 27   esriGeometryPolygon
#> 28   esriGeometryPolygon
#> 29   esriGeometryPolygon
#> 30   esriGeometryPolygon
#> 31   esriGeometryPolygon
#> 32   esriGeometryPolygon
#> 33   esriGeometryPolygon
#> 34   esriGeometryPolygon
#> 35   esriGeometryPolygon
#> 36   esriGeometryPolygon
#> 37   esriGeometryPolygon
#> 38   esriGeometryPolygon
#> 39   esriGeometryPolygon
#> 40   esriGeometryPolygon
#> 41   esriGeometryPolygon
#> 42   esriGeometryPolygon
#> 43   esriGeometryPolygon
#> 44   esriGeometryPolygon
#> 45   esriGeometryPolygon
#> 46   esriGeometryPolygon
#> 47   esriGeometryPolygon
#> 48   esriGeometryPolygon
#> 49  esriGeometryPolyline
#> 50   esriGeometryPolygon
#> 51   esriGeometryPolygon
#> 52   esriGeometryPolygon
#> 53   esriGeometryPolygon
#> 54   esriGeometryPolygon
#> 55   esriGeometryPolygon
#> 56   esriGeometryPolygon
#> 57   esriGeometryPolygon
#> 58   esriGeometryPolygon
#> 59   esriGeometryPolygon
#> 60   esriGeometryPolygon
#> 61  esriGeometryPolyline
#> 62   esriGeometryPolygon
#> 63   esriGeometryPolygon
#> 64   esriGeometryPolygon
#> 65   esriGeometryPolygon
#> 66   esriGeometryPolygon
#> 67   esriGeometryPolygon
#> 68   esriGeometryPolygon
#> 69   esriGeometryPolygon
#> 70   esriGeometryPolygon
#> 71   esriGeometryPolygon
#> 72   esriGeometryPolygon
#> 73   esriGeometryPolygon
#> 74   esriGeometryPolygon
#> 75   esriGeometryPolygon
#> 76   esriGeometryPolygon
#> 77   esriGeometryPolygon
#> 78   esriGeometryPolygon
#> 79   esriGeometryPolygon
#> 80   esriGeometryPolygon
#> 81   esriGeometryPolygon
#> 82   esriGeometryPolygon
#> 83   esriGeometryPolygon
#> 84   esriGeometryPolygon
#> 85   esriGeometryPolygon
#> 86   esriGeometryPolygon
#> 87   esriGeometryPolygon
#> 88   esriGeometryPolygon
#> 89   esriGeometryPolygon
#> 90   esriGeometryPolygon
#> 91   esriGeometryPolygon
#> 92   esriGeometryPolygon
#> 93   esriGeometryPolygon
#> 94   esriGeometryPolygon
#> 95   esriGeometryPolygon
#> 96   esriGeometryPolygon
#> 97   esriGeometryPolygon
#> 98   esriGeometryPolygon
#> 99   esriGeometryPolygon
#> 100  esriGeometryPolygon
#> 101  esriGeometryPolygon
#> 102  esriGeometryPolygon
#> 103  esriGeometryPolygon
#> 104  esriGeometryPolygon
#> 105    esriGeometryPoint
#> 106  esriGeometryPolygon
#> 107 esriGeometryPolyline
#> 108 esriGeometryPolyline
#> 109  esriGeometryPolygon
#> 110  esriGeometryPolygon
#> 111  esriGeometryPolygon
#> 112  esriGeometryPolygon
#> 113  esriGeometryPolygon
#> 114  esriGeometryPolygon
#> 115  esriGeometryPolygon
#> 116  esriGeometryPolygon
#> 117  esriGeometryPolygon
#> 118  esriGeometryPolygon
#> 119  esriGeometryPolygon
#> 120  esriGeometryPolygon
#> 121  esriGeometryPolygon
#> 122  esriGeometryPolygon
#> 123  esriGeometryPolygon
#> 124  esriGeometryPolygon
#> 125  esriGeometryPolygon
#> 126    esriGeometryPoint
#> 127  esriGeometryPolygon
#> 128    esriGeometryPoint
#> 129  esriGeometryPolygon
#> 130  esriGeometryPolygon
#> 131  esriGeometryPolygon
#> 132  esriGeometryPolygon
#> 133  esriGeometryPolygon
#> 134    esriGeometryPoint
#> 135  esriGeometryPolygon
#> 136  esriGeometryPolygon
#> 137    esriGeometryPoint
#> 138  esriGeometryPolygon
#> 139  esriGeometryPolygon
#> 140 esriGeometryPolyline
#> 141  esriGeometryPolygon
#> 142    esriGeometryPoint
#> 143    esriGeometryPoint
#> 144  esriGeometryPolygon
#> 145    esriGeometryPoint
#> 146    esriGeometryPoint
#> 147  esriGeometryPolygon
#> 148  esriGeometryPolygon
#> 149    esriGeometryPoint
#> 150  esriGeometryPolygon
#> 151  esriGeometryPolygon
#> 152  esriGeometryPolygon
#> 153  esriGeometryPolygon
#> 154  esriGeometryPolygon
#> 155  esriGeometryPolygon
#> 156  esriGeometryPolygon
#> 157    esriGeometryPoint
#> 158  esriGeometryPolygon
#> 159    esriGeometryPoint
#> 160  esriGeometryPolygon
#> 161  esriGeometryPolygon
#> 162  esriGeometryPolygon
#> 163  esriGeometryPolygon
#> 164    esriGeometryPoint
#> 165  esriGeometryPolygon
#> 166 esriGeometryPolyline
#> 167  esriGeometryPolygon
#> 168  esriGeometryPolygon
#> 169  esriGeometryPolygon
#> 170  esriGeometryPolygon
#> 171  esriGeometryPolygon
#> 172  esriGeometryPolygon
#> 173  esriGeometryPolygon
#> 174  esriGeometryPolygon
#> 175  esriGeometryPolygon
#> 176  esriGeometryPolygon
#> 177    esriGeometryPoint
#> 178    esriGeometryPoint
#> 179 esriGeometryPolyline
#> 180 esriGeometryPolyline
#> 181    esriGeometryPoint
#> 182    esriGeometryPoint
#> 183  esriGeometryPolygon
#> 184  esriGeometryPolygon
#> 185    esriGeometryPoint
#> 186  esriGeometryPolygon
#> 187  esriGeometryPolygon
#> 188 esriGeometryPolyline
#> 189  esriGeometryPolygon
#> 190  esriGeometryPolygon
#> 191  esriGeometryPolygon
#> 192    esriGeometryPoint
#> 193  esriGeometryPolygon
#> 194    esriGeometryPoint
#> 195 esriGeometryPolyline
#> 196  esriGeometryPolygon
#> 197  esriGeometryPolygon
#> 198  esriGeometryPolygon
#> 199  esriGeometryPolygon
#> 200  esriGeometryPolygon
#> 201 esriGeometryPolyline
#> 202    esriGeometryPoint
#> 203    esriGeometryPoint
#> 204    esriGeometryPoint
#> 205    esriGeometryPoint
#> 206 esriGeometryPolyline
#> 207 esriGeometryPolyline
#> 208    esriGeometryPoint
#> 209 esriGeometryPolyline
#> 210 esriGeometryPolyline
#> 211  esriGeometryPolygon
#> 212 esriGeometryPolyline
#> 213    esriGeometryPoint
#> 214 esriGeometryPolyline
#> 215    esriGeometryPoint
#> 216 esriGeometryPolyline
#> 217    esriGeometryPoint
#> 218 esriGeometryPolyline
#> 219 esriGeometryPolyline
#> 220 esriGeometryPolyline
#> 221    esriGeometryPoint
#> 222    esriGeometryPoint
#> 223    esriGeometryPoint
#> 224 esriGeometryPolyline
#> 225 esriGeometryPolyline
#> 226    esriGeometryPoint
#> 227  esriGeometryPolygon
#> 228    esriGeometryPoint
#> 229 esriGeometryPolyline
#> 230 esriGeometryPolyline
#> 231 esriGeometryPolyline
#> 232    esriGeometryPoint
#> 233    esriGeometryPoint
#> 234 esriGeometryPolyline
#> 235 esriGeometryPolyline
#> 236 esriGeometryPolyline
#> 237 esriGeometryPolyline
#> 238 esriGeometryPolyline
#> 239    esriGeometryPoint
#> 240    esriGeometryPoint
#> 241    esriGeometryPoint
#> 242 esriGeometryPolyline
#> 243 esriGeometryPolyline
#> 244 esriGeometryPolyline
#> 245 esriGeometryPolyline
#> 246 esriGeometryPolyline
#> 247    esriGeometryPoint
#> 248 esriGeometryPolyline
#> 249 esriGeometryPolyline
#> 250 esriGeometryPolyline
#> 251 esriGeometryPolyline
#> 252 esriGeometryPolyline
#> 253    esriGeometryPoint
#> 254 esriGeometryPolyline
#> 255    esriGeometryPoint
#> 256 esriGeometryPolyline
#> 257 esriGeometryPolyline
#> 258 esriGeometryPolyline
#> 259  esriGeometryPolygon
#> 260    esriGeometryPoint
#> 261  esriGeometryPolygon
#> 262  esriGeometryPolygon
#> 263  esriGeometryPolygon
#> 264    esriGeometryPoint
#> 265 esriGeometryPolyline
#> 266  esriGeometryPolygon
#> 267 esriGeometryPolyline
#> 268 esriGeometryPolyline
#> 269 esriGeometryPolyline
#> 270 esriGeometryPolyline
#> 271    esriGeometryPoint
#> 272    esriGeometryPoint
#> 273    esriGeometryPoint
#> 274  esriGeometryPolygon
#> 275    esriGeometryPoint
#> 276 esriGeometryPolyline
#> 277  esriGeometryPolygon
#> 278  esriGeometryPolygon
#> 279  esriGeometryPolygon
#> 280  esriGeometryPolygon
#> 281  esriGeometryPolygon
#> 282  esriGeometryPolygon
#> 283  esriGeometryPolygon
#> 284  esriGeometryPolygon
#> 285  esriGeometryPolygon
list_feature_classes(feature_dataset = "equity")
#>               layer_name feature_dataset       geometry_type
#> 1      DISPLACEMENT_2018          equity esriGeometryPolygon
#> 2      DISPLACEMENT_RISK          equity esriGeometryPolygon
#> 3 DISPLACEMENT_RISK_MASK          equity esriGeometryPolygon
#> 4          EQUITY_TRACTS          equity esriGeometryPolygon
#> 5       OPPORTUNITY_MASK          equity esriGeometryPolygon
#> 6        REDLINING_AREAS          equity esriGeometryPolygon