list_recordsets(schema_name, recordset_type)

return a list of tables and views available in PSRC's in-house data warehouse

Usage

list_recordsets(schema_name = "", include_base_tables = FALSE)

Arguments

schema_name

(string, optional) If supplied, this limits the returned list to only those tables and views in schema_name.

include_base_tables

(TRUE|FALSE, optional) If false, return only the views (no tables). False is the default.

Value

a list of tables, in the form of a dataframe

Examples

list_recordsets()
#>                  schema
#> 1           small_areas
#> 2              HHSurvey
#> 3                equity
#> 4                equity
#> 5                equity
#> 6              HHSurvey
#> 7         regional_gtfs
#> 8         regional_gtfs
#> 9         regional_gtfs
#> 10        regional_gtfs
#> 11        regional_gtfs
#> 12               census
#> 13        regional_gtfs
#> 14        regional_gtfs
#> 15               census
#> 16 employment_summaries
#> 17               census
#> 18 employment_summaries
#> 19               census
#> 20               census
#> 21 employment_summaries
#> 22                  ofm
#> 23                  ofm
#> 24           employment
#> 25                  ofm
#> 26            Political
#> 27    parking_inventory
#> 28          data_portal
#> 29               census
#> 30            ofm_views
#> 31       seatac_airport
#> 32             HHSurvey
#> 33        Macroeconomic
#> 34        Macroeconomic
#> 35        Macroeconomic
#> 36        Macroeconomic
#> 37                 chas
#> 38             HHSurvey
#> 39               equity
#> 40           employment
#> 41             HHSurvey
#> 42        park_and_ride
#> 43    parking_inventory
#> 44               census
#> 45                  LUV
#> 46          small_areas
#> 47          small_areas
#> 48          small_areas
#> 49                  LUV
#> 50           employment
#> 51          small_areas
#> 52          small_areas
#> 53                  LUV
#> 54          small_areas
#> 55                  LUV
#> 56          small_areas
#> 57                  LUV
#> 58                  LUV
#> 59                  LUV
#> 60             HHSurvey
#> 61             HHSurvey
#> 62             HHSurvey
#>                                              recordset_name recordset_type
#> 1                                            v_parcels_2018           VIEW
#> 2                                          v_value_metadata           VIEW
#> 3                                       v_blockgroup_shares           VIEW
#> 4                                      v_tracker_indicators           VIEW
#> 5                                            v_tract_shares           VIEW
#> 6                                                 v_persons           VIEW
#> 7                                             v_agency_2018           VIEW
#> 8                                           v_calendar_2018           VIEW
#> 9                                             v_routes_2018           VIEW
#> 10                                        v_service_id_2018           VIEW
#> 11                                        v_stop_times_2018           VIEW
#> 12                                          v_geography_dim           VIEW
#> 13                                             v_stops_2018           VIEW
#> 14                                             v_trips_2018           VIEW
#> 15                               v_data_for_member_profiles           VIEW
#> 16                   v_covered_employment_by_city_by_naics2           VIEW
#> 17                               v_geo_relationships_blocks           VIEW
#> 18             v_covered_employment_by_city_by_major_sector           VIEW
#> 19                         v_geo_relationships_block_groups           VIEW
#> 20                               v_geo_relationships_tracts           VIEW
#> 21            v_covered_employment_by_tract_by_major_sector           VIEW
#> 22                                         v_estimates_2019           VIEW
#> 23                                v_current_block_estimates           VIEW
#> 24                                                test_view           VIEW
#> 25                                         v_estimates_2020           VIEW
#> 26                                   incorporated_city_dims           VIEW
#> 27                                 v_parking_inventory_view           VIEW
#> 28                                 v_parking_by_block_group           VIEW
#> 29                                      v_pums_persons_2019           VIEW
#> 30                                         v_estimates_2019           VIEW
#> 31                                           v_measurements           VIEW
#> 32                              v_column_definitions_public           VIEW
#> 33                                             v_employment           VIEW
#> 34                                             v_households           VIEW
#> 35                                                 v_income           VIEW
#> 36                                                    v_pop           VIEW
#> 37 v_households_by_cost_burden_by_tenure_by_race_by_subarea           VIEW
#> 38                                         v_trips_in_house           VIEW
#> 39                                      v_blockgroup_counts           VIEW
#> 40                           v_hct_station_areas_employment           VIEW
#> 41                                             v_households           VIEW
#> 42                                         v_park_and_rides           VIEW
#> 43                                              v_test_view           VIEW
#> 44  v_pums_households_by_rent_cost_burden_by_income_by_race           VIEW
#> 45                               v_jobs_by_sector_by_county           VIEW
#> 46                                        v_household_facts           VIEW
#> 47                               v_jobs_by_sector_by_county           VIEW
#> 48                                              v_job_facts           VIEW
#> 49                                                   v_jobs           VIEW
#> 50                                         custom_geography           VIEW
#> 51                                                   v_jobs           VIEW
#> 52                                            v_parcel_dims           VIEW
#> 53                                        v_household_facts           VIEW
#> 54                                       v_population_facts           VIEW
#> 55                                              v_job_facts           VIEW
#> 56                                            v_sector_dims           VIEW
#> 57                                            v_parcel_dims           VIEW
#> 58                                       v_population_facts           VIEW
#> 59                                            v_sector_dims           VIEW
#> 60                                                  v_trips           VIEW
#> 61                                                   v_days           VIEW
#> 62                                               v_vehicles           VIEW

list_recordsets(schema_name="HHSurvey")
#>     schema              recordset_name recordset_type
#> 1 HHSurvey v_column_definitions_public           VIEW
#> 2 HHSurvey                      v_days           VIEW
#> 3 HHSurvey                v_households           VIEW
#> 4 HHSurvey                   v_persons           VIEW
#> 5 HHSurvey                     v_trips           VIEW
#> 6 HHSurvey            v_trips_in_house           VIEW
#> 7 HHSurvey            v_value_metadata           VIEW
#> 8 HHSurvey                  v_vehicles           VIEW

list_recordsets(schema_name="HHSurvey", include_base_tables=TRUE)
#>      schema              recordset_name recordset_type
#> 1  HHSurvey              days_2017_2019     BASE TABLE
#> 2  HHSurvey                   days_2021     BASE TABLE
#> 3  HHSurvey             hh_survey_audit     BASE TABLE
#> 4  HHSurvey          household_dim_2014     BASE TABLE
#> 5  HHSurvey          household_dim_2017     BASE TABLE
#> 6  HHSurvey         household_fact_2015     BASE TABLE
#> 7  HHSurvey        households_2017_2019     BASE TABLE
#> 8  HHSurvey             households_2021     BASE TABLE
#> 9  HHSurvey             person_dim_2014     BASE TABLE
#> 10 HHSurvey             person_dim_2015     BASE TABLE
#> 11 HHSurvey            person_fact_2014     BASE TABLE
#> 12 HHSurvey            person_fact_2015     BASE TABLE
#> 13 HHSurvey           persons_2017_2019     BASE TABLE
#> 14 HHSurvey                persons_2021     BASE TABLE
#> 15 HHSurvey               trip_dim_2014     BASE TABLE
#> 16 HHSurvey               trip_dim_2015     BASE TABLE
#> 17 HHSurvey              trip_fact_2014     BASE TABLE
#> 18 HHSurvey             trips_2017_2019     BASE TABLE
#> 19 HHSurvey                  trips_2021     BASE TABLE
#> 20 HHSurvey v_column_definitions_public           VIEW
#> 21 HHSurvey                      v_days           VIEW
#> 22 HHSurvey                v_households           VIEW
#> 23 HHSurvey                   v_persons           VIEW
#> 24 HHSurvey                     v_trips           VIEW
#> 25 HHSurvey            v_trips_in_house           VIEW
#> 26 HHSurvey            v_value_metadata           VIEW
#> 27 HHSurvey                  v_vehicles           VIEW
#> 28 HHSurvey              value_metadata     BASE TABLE
#> 29 HHSurvey           variable_metadata     BASE TABLE
#> 30 HHSurvey          variable_metadata2     BASE TABLE
#> 31 HHSurvey          variables_codebook     BASE TABLE
#> 32 HHSurvey          vehicles_2017_2019     BASE TABLE
#> 33 HHSurvey               vehicles_2021     BASE TABLE