st_read_elmergeo(layer_name, ...)

read simple features from PSRC's in-house geodatabase #' This function is preferable to sf::st_read() when pulling from Elmer or ElmerGeo, because it avoids some incompatibilities between SQL Server's data types and R's.

Usage

st_read_elmergeo(layer_name, schema_name = "dbo", project_to_wgs84 = TRUE)

Arguments

layer_name

The name of the feature layer or geodatabase table

schema_name

The name of the schema that layer_name exists in. Defaults to "dbo", the standard schema for ElmerGeo tables

project_to_wgs84

(TRUE/FALSE) If TRUE then deliver the output in WGS84 projection, otherwise NAD84 / WA State Plane North. Defaults to TRUE.

Value

object of class sf

Note

If the layer has been set up as an ESRI versioned layer in the geodatabase, this function returns the versioned view (which exists in SQL Server with a "_evw" suffix). If it has not been set up that way, it returns the base table.

Examples

st_read_elmergeo("COUNTY_BACKGROUND")
#> Simple feature collection with 39 features and 7 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -124.8363 ymin: 45.54373 xmax: -116.9152 ymax: 49.00253
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    OBJECTID perimeter county_cod county_fip county_nm psrc SDE_STATE_ID
#> 1        23 1287929.0         38        075   Whitman    0            0
#> 2         3 1291742.0         10        019     Ferry    0            0
#> 3        26 1702496.0         39        077    Yakima    0            0
#> 4         6  601010.2         28        055  San Juan    0            0
#> 5        29  962960.1         11        021  Franklin    0            0
#> 6         9 1788454.0          4        007    Chelan    0            0
#> 7        15  631297.4         18        035    Kitsap    1            0
#> 8        32  796659.6          7        013  Columbia    0            0
#> 9        12 1197390.0         31        061 Snohomish    1            0
#> 10       35  416193.8         35        069 Wahkiakum    0            0
#>                             Shape
#> 1  POLYGON ((-117.0403 47.2592...
#> 2  POLYGON ((-118.8366 49.0001...
#> 3  POLYGON ((-121.4076 47.0006...
#> 4  POLYGON ((-123.0084 48.8296...
#> 5  POLYGON ((-118.2102 46.7388...
#> 6  POLYGON ((-120.7021 48.5317...
#> 7  POLYGON ((-122.4983 47.9266...
#> 8  POLYGON ((-118.2286 46.5933...
#> 9  POLYGON ((-122.4583 48.2977...
#> 10 POLYGON ((-123.2184 46.3850...

st_read_elmergeo("COUNTY_BACKGROUND", project_to_wgs84 = FALSE)
#> Simple feature collection with 39 features and 7 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 664166.8 ymin: -527832.4 xmax: 2636448 ymax: 751796.4
#> Projected CRS: NAD83 / Washington North (ftUS)
#> First 10 features:
#>    OBJECTID perimeter county_cod county_fip county_nm psrc SDE_STATE_ID
#> 1        23 1287929.0         38        075   Whitman    0            0
#> 2         3 1291742.0         10        019     Ferry    0            0
#> 3        26 1702496.0         39        077    Yakima    0            0
#> 4         6  601010.2         28        055  San Juan    0            0
#> 5        29  962960.1         11        021  Franklin    0            0
#> 6         9 1788454.0          4        007    Chelan    0            0
#> 7        15  631297.4         18        035    Kitsap    1            0
#> 8        32  796659.6          7        013  Columbia    0            0
#> 9        12 1197390.0         31        061 Snohomish    1            0
#> 10       35  416193.8         35        069 Wahkiakum    0            0
#>                             Shape
#> 1  POLYGON ((2581969 117791.3,...
#> 2  POLYGON ((2119733 735853.9,...
#> 3  POLYGON ((1497107 772.0603,...
#> 4  POLYGON ((1116546 674848.3,...
#> 5  POLYGON ((2298154 -84060.12...
#> 6  POLYGON ((1672219 558791.7,...
#> 7  POLYGON ((1232246 342417.9,...
#> 8  POLYGON ((2295340 -137255.1...
#> 9  POLYGON ((1244905 477543.5,...
#> 10 POLYGON ((1038351 -214998.9...